Moved functions to a utils file.
